refactor: Dashboard component📦
